requirements.txt
Django>=3.1.7,<3.2.0
djangorestframework>=3.11.2,<3.12.0
drf-social-oauth2>=1.0.9,<1.1.0
django-cors-headers>=3.7.0,<3.8.0
django-use-email-as-username>=1.1.2,<1.2.0
psycopg2>=2.8.6,<2.9.0
flake8>=3.8.4,<3.9.0
